Extract error screen to a separate file
